backend/pyproject.toml
[project]
name = "exam-backend"
version = "0.1.0"
description = "Online Exam Management System Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "psycopg[binary]>=3.1.13",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "email-validator>=2.1.0",
    "pytest>=9.0.1",
    "httpx>=0.28.1",
    "alembic>=1.17.2",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# --- ADD THIS SECTION ---
[tool.setuptools]
packages = ["app"]
